\name{FunkyTax-package}
\alias{FunkyTax-package}
\alias{FunkyTax}
\docType{package}
\title{
\packageTitle{FunkyTax}
}
\description{
\packageDescription{FunkyTax}
}
\details{

Use TaFuR to parse a matrix of functions by OTUs by samples into individual community matrices for each function.  Resulting matrices include only OTUs contributing to the respective function. TaFuR also performs Hellinger transformations (Legendre and Gallagher 2001) on the matrices, computes corresponding distance matrices, and performs ADONIS.  Use CatFun to compare these results with those from DESeq2 (Love et al. 2014) to classify functions and community subsets into functional community categories.
}
\author{
\packageAuthor{FunkyTax}

Maintainer: \packageMaintainer{FunkyTax}
}
\references{
Anderson, MJ (2001) A new method for non-parametric multivariate analysis of variance, \emph{Austral Ecology}, 26: 32–46.

Legendre P, Gallagher ED (2001) Ecologically meaningful transformation ofr ordication of species data, \emph{Oecologia}. 129:271-280.

Love MI, Huber W, Anders S (2014) Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2, \emph{Genome Biology}, 15:550.

Phillips CD, Hanson J, Wilkinson J, Koenig L, Rees E, Webala P, Kingston T (submitted) Microbiome structural and functional incongruence explained along host dietary niche space.

}

\keyword{ package }
